Computational Solid Geometry

A solid modeler for Julia.

A package for building 2D/3D solid models by Boolean combination of solid primitives.

Note

Please read through the LAR Documentation section of the LAR modeling pages if this is your first time using LAR software. Then come back here.

Note that the above must be updated to address the Model concept introduced here.

Package Features

Algebraic approach to computate Boolean operations between solid models

Boolean forms evaluated assembling atoms of finite algebra generated by $E^d$ arrangement of solid terms.

Atoms (irreducible 3-cells) give a basis of the linear chain space associated with this partition.

Basic linear algebra and algebraic topology is used, namely sparse matrices of linear operators.

Validity of topology computations is guaranteed.

CSG expressions of arbitrary complexity evaluated by a novel efficient process.

CSG forms of any complexity traduced to BitArrays array, evaluated by vectorized logical operations.

Distinction is removed between manifold and non-manifold models.

Simple DSL (Domain Specific Language) under development via Julia macros.

<!–

The Package Guide provides a tutorial explaining how to get started using Documenter.

Some examples of packages using Documenter can be found on the Examples page.

See the Index for the complete list of documented functions and types.

Manual Outline

Library Outline

Index

–>